Permisos en el reporte
Se elimino el fragment del reporte migrando todo al activity de reporte y se agregaron los permisos al manifest
