+++ .idea/misc.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="..\:/edith/Documents/GitHub/TamayoReport/app/src/main/res/layout/activity_categories.xml" value="0.3723958333333333" />
+        <entry key="..\:/edith/Documents/GitHub/TamayoReport/app/src/main/res/layout/activity_reporte.xml" value="0.3723958333333333" />
+        <entry key="..\:/edith/Documents/GitHub/TamayoReport/app/src/main/res/layout/fragment_reporte.xml" value="0.3723958333333333" />
+      </map>
+    </option>
+  </component>
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+    <output url="file://$PROJECT_DIR$/build/classes" />
+  </component>
+  <component name="ProjectType">
+    <option name="id" value="Android" />
+  </component>
+</project>
