+++ app/src/main/res/layout/activity_categories.xml
@@ -59,8 +59,8 @@
         android:layout_marginTop="16dp"
         android:background="@color/aqua"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView5">
-
+        app:layout_constraintTop_toBottomOf="@+id/textView5"
+        android:radius="1000dp">
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
