+++ app/src/main/res/layout/activity_registro.xml
@@ -113,10 +113,10 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.501"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/categoriaReporte" />
+        app:layout_constraintTop_toBottomOf="@+id/reportCategory" />
 
     <TextView
-        android:id="@+id/categoriaReporte"
+        android:id="@+id/reportCategory"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="40dp"
